--- a/requirements.txt
+++ b/requirements.txt
@@ -1,6 +1,6 @@
-jax==0.2.9
-jaxlib==0.1.60
-objax==1.3.1
+jax==0.4.2
+jaxlib==0.4.2
+objax==1.6.0
 numba
 numpy
 matplotlib