We have been stuck with `jax==0.2.9`, `jaxlib==0.1.60` and `objax==1.3.1` for some years now due to #3 where updating objax/jax caused the model to be compiled twice leading to slowdown. This seems to have been fixed in the most recent jax/objax versions and thus updating Bayes-Newton to use more recent versions of them is sensible.

FIXES #3

Changes due to updates in jax:
* The functions jax.ops.index_update, jax.ops.index_add, which were deprecated in 0.2.22, have been removed. Please use [the .at property on JAX arrays](https://jax.readthedocs.io/en/latest/_autosummary/jax.numpy.ndarray.at.html) instead, e.g., x.at[idx].set(y).