@asolin 2023-02-07
The new version seems to be marginally faster than the old---at least in most cases: Results on sparse GP classification:
```
OLD:
Training Sparse-VI
iter 10, energy: 207.1191
iter 20, energy: 176.0094
iter 30, energy: 167.5217
iter 40, energy: 163.0514
iter 50, energy: 159.5225
iter 60, energy: 157.6291
iter 70, energy: 156.1543
iter 80, energy: 155.1126
iter 90, energy: 154.2804
iter 100, energy: 153.5996
iter 110, energy: 153.0288
iter 120, energy: 152.5357
iter 130, energy: 152.1057
iter 140, energy: 151.7247
iter 150, energy: 151.3838
iter 160, energy: 151.0762
iter 170, energy: 150.7966
iter 180, energy: 150.5407
iter 190, energy: 150.3055
iter 200, energy: 150.0881
Total training time: 11.85 seconds
NLPD: 0.2535
Training Sparse-EP
iter 10, energy: 174.0703
iter 20, energy: 136.5381
iter 30, energy: 123.5028
iter 40, energy: 117.4766
iter 50, energy: 113.0087
iter 60, energy: 110.1356
iter 70, energy: 108.6130
iter 80, energy: 106.8834
iter 90, energy: 106.0489
iter 100, energy: 105.1179
iter 110, energy: 104.4026
iter 120, energy: 103.8325
iter 130, energy: 103.2922
iter 140, energy: 102.8285
iter 150, energy: 102.4203
iter 160, energy: 102.0456
iter 170, energy: 101.7052
iter 180, energy: 101.3954
iter 190, energy: 101.1099
iter 200, energy: 100.8457
Total training time: 11.26 seconds
NLPD: 0.2671
Training Sparse-LA
iter 10, energy: 184.1977
iter 20, energy: 150.6769
iter 30, energy: 141.9323
iter 40, energy: 136.3155
iter 50, energy: 131.4491
iter 60, energy: 129.0010
iter 70, energy: 126.5017
iter 80, energy: 124.9792
iter 90, energy: 123.4974
iter 100, energy: 122.4107
iter 110, energy: 121.3903
iter 120, energy: 120.5535
iter 130, energy: 119.7800
iter 140, energy: 119.1040
iter 150, energy: 118.4821
iter 160, energy: 117.9186
iter 170, energy: 117.3992
iter 180, energy: 116.9191
iter 190, energy: 116.4738
iter 200, energy: 116.0579
Total training time: 8.22 seconds
NLPD: 0.2542
```
<img width="790" alt="image" src="https://user-images.githubusercontent.com/2668346/217198463-ca61e85e-695b-4573-8bb3-579b9a9180e8.png">


```
NEW:
Training Sparse-VI
iter 10, energy: 207.1191
iter 20, energy: 176.0094
iter 30, energy: 167.5217
iter 40, energy: 163.0514
iter 50, energy: 159.5225
iter 60, energy: 157.6291
iter 70, energy: 156.1543
iter 80, energy: 155.1126
iter 90, energy: 154.2804
iter 100, energy: 153.5996
iter 110, energy: 153.0288
iter 120, energy: 152.5357
iter 130, energy: 152.1057
iter 140, energy: 151.7247
iter 150, energy: 151.3838
iter 160, energy: 151.0762
iter 170, energy: 150.7966
iter 180, energy: 150.5407
iter 190, energy: 150.3055
iter 200, energy: 150.0881
Total training time: 6.36 seconds
NLPD: 0.2535
Training Sparse-EP
iter 10, energy: 174.0703
iter 20, energy: 136.5381
iter 30, energy: 123.5028
iter 40, energy: 117.4766
iter 50, energy: 113.0087
iter 60, energy: 110.1356
iter 70, energy: 108.6130
iter 80, energy: 106.8834
iter 90, energy: 106.0489
iter 100, energy: 105.1179
iter 110, energy: 104.4026
iter 120, energy: 103.8325
iter 130, energy: 103.2922
iter 140, energy: 102.8285
iter 150, energy: 102.4203
iter 160, energy: 102.0456
iter 170, energy: 101.7052
iter 180, energy: 101.3954
iter 190, energy: 101.1099
iter 200, energy: 100.8457
Total training time: 7.40 seconds
NLPD: 0.2671
Training Sparse-LA
iter 10, energy: 184.1977
iter 20, energy: 150.6769
iter 30, energy: 141.9323
iter 40, energy: 136.3155
iter 50, energy: 131.4491
iter 60, energy: 129.0010
iter 70, energy: 126.5017
iter 80, energy: 124.9792
iter 90, energy: 123.4974
iter 100, energy: 122.4107
iter 110, energy: 121.3903
iter 120, energy: 120.5535
iter 130, energy: 119.7800
iter 140, energy: 119.1040
iter 150, energy: 118.4821
iter 160, energy: 117.9186
iter 170, energy: 117.3992
iter 180, energy: 116.9191
iter 190, energy: 116.4738
iter 200, energy: 116.0579
Total training time: 4.52 seconds
NLPD: 0.2542
```
<img width="787" alt="image" src="https://user-images.githubusercontent.com/2668346/217198397-1219bc32-9cea-4667-bd35-b7de07b528fe.png">